Uncommented looping Quick Start Guide
Took out the comments on the looping section and changed the looping to every 3 seconds for 5 minutes.  This gives the user the chance to check their work in the developer view, gives them enough samples to get a foghorn alert and allows them to see views as they add new data.